разбиение джс на файлы модулей
